+++ internal/balancer/connections_state_test.go
@@ -1,4 +1,4 @@
-package router
+package balancer
 
 import (
 	"context"
@@ -7,7 +7,7 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	routerconfig "github.com/ydb-platform/ydb-go-sdk/v3/internal/router/config"
+	balancerConfig "github.com/ydb-platform/ydb-go-sdk/v3/internal/balancer/config"
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/conn"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/mock"
@@ -68,7 +68,7 @@ func TestSortPreferConnections(t *testing.T) {
 		name          string
 		source        []conn.Conn
 		allowFallback bool
-		filter        routerconfig.PreferConnFunc
+		filter        balancerConfig.PreferConnFunc
 		prefer        []conn.Conn
 		fallback      []conn.Conn
 	}{
@@ -103,7 +103,7 @@ func TestSortPreferConnections(t *testing.T) {
 				&mock.Conn{AddrField: "f2"},
 			},
 			allowFallback: false,
-			filter: func(_ routerconfig.Info, c conn.Conn) bool {
+			filter: func(_ balancerConfig.Info, c conn.Conn) bool {
 				return strings.HasPrefix(c.Endpoint().Address(), "t")
 			},
 			prefer: []conn.Conn{
@@ -121,7 +121,7 @@ func TestSortPreferConnections(t *testing.T) {
 				&mock.Conn{AddrField: "f2"},
 			},
 			allowFallback: true,
-			filter: func(_ routerconfig.Info, c conn.Conn) bool {
+			filter: func(_ balancerConfig.Info, c conn.Conn) bool {
 				return strings.HasPrefix(c.Endpoint().Address(), "t")
 			},
 			prefer: []conn.Conn{
@@ -137,7 +137,7 @@ func TestSortPreferConnections(t *testing.T) {
 
 	for _, test := range table {
 		t.Run(test.name, func(t *testing.T) {
-			prefer, fallback := sortPreferConnections(test.source, test.filter, routerconfig.Info{}, test.allowFallback)
+			prefer, fallback := sortPreferConnections(test.source, test.filter, balancerConfig.Info{}, test.allowFallback)
 			require.Equal(t, test.prefer, prefer)
 			require.Equal(t, test.fallback, fallback)
 		})
@@ -145,7 +145,7 @@ func TestSortPreferConnections(t *testing.T) {
 }
 
 func TestSelectRandomConnection(t *testing.T) {
-	s := newConnectionsState(nil, nil, routerconfig.Info{}, false)
+	s := newConnectionsState(nil, nil, balancerConfig.Info{}, false)
 
 	t.Run("Empty", func(t *testing.T) {
 		c, failedCount := s.selectRandomConnection(nil, false)
@@ -237,7 +237,7 @@ func TestNewState(t *testing.T) {
 	}{
 		{
 			name:  "Empty",
-			state: newConnectionsState(nil, nil, routerconfig.Info{}, false),
+			state: newConnectionsState(nil, nil, balancerConfig.Info{}, false),
 			res: &connectionsState{
 				connByNodeID: nil,
 				prefer:       nil,
@@ -250,7 +250,7 @@ func TestNewState(t *testing.T) {
 			state: newConnectionsState([]conn.Conn{
 				&mock.Conn{AddrField: "1"},
 				&mock.Conn{AddrField: "2"},
-			}, nil, routerconfig.Info{}, false),
+			}, nil, balancerConfig.Info{}, false),
 			res: &connectionsState{
 				connByNodeID: nil,
 				prefer: []conn.Conn{
@@ -271,9 +271,9 @@ func TestNewState(t *testing.T) {
 				&mock.Conn{AddrField: "f1", LocationField: "f"},
 				&mock.Conn{AddrField: "t2", LocationField: "t"},
 				&mock.Conn{AddrField: "f2", LocationField: "f"},
-			}, func(routerInfo routerconfig.Info, c conn.Conn) bool {
-				return routerInfo.SelfLocation == c.Endpoint().Location()
-			}, routerconfig.Info{SelfLocation: "t"}, false),
+			}, func(info balancerConfig.Info, c conn.Conn) bool {
+				return info.SelfLocation == c.Endpoint().Location()
+			}, balancerConfig.Info{SelfLocation: "t"}, false),
 			res: &connectionsState{
 				connByNodeID: nil,
 				prefer: []conn.Conn{
@@ -294,9 +294,9 @@ func TestNewState(t *testing.T) {
 				&mock.Conn{AddrField: "f1", LocationField: "f"},
 				&mock.Conn{AddrField: "t2", LocationField: "t"},
 				&mock.Conn{AddrField: "f2", LocationField: "f"},
-			}, func(routerInfo routerconfig.Info, c conn.Conn) bool {
-				return routerInfo.SelfLocation == c.Endpoint().Location()
-			}, routerconfig.Info{SelfLocation: "t"}, true),
+			}, func(info balancerConfig.Info, c conn.Conn) bool {
+				return info.SelfLocation == c.Endpoint().Location()
+			}, balancerConfig.Info{SelfLocation: "t"}, true),
 			res: &connectionsState{
 				connByNodeID: nil,
 				prefer: []conn.Conn{
@@ -322,9 +322,9 @@ func TestNewState(t *testing.T) {
 				&mock.Conn{AddrField: "f1", NodeIDField: 2, LocationField: "f"},
 				&mock.Conn{AddrField: "t2", NodeIDField: 3, LocationField: "t"},
 				&mock.Conn{AddrField: "f2", NodeIDField: 4, LocationField: "f"},
-			}, func(routerInfo routerconfig.Info, c conn.Conn) bool {
-				return routerInfo.SelfLocation == c.Endpoint().Location()
-			}, routerconfig.Info{SelfLocation: "t"}, true),
+			}, func(info balancerConfig.Info, c conn.Conn) bool {
+				return info.SelfLocation == c.Endpoint().Location()
+			}, balancerConfig.Info{SelfLocation: "t"}, true),
 			res: &connectionsState{
 				connByNodeID: map[uint32]conn.Conn{
 					1: &mock.Conn{AddrField: "t1", NodeIDField: 1, LocationField: "t"},
@@ -361,7 +361,7 @@ func TestNewState(t *testing.T) {
 
 func TestConnection(t *testing.T) {
 	t.Run("Empty", func(t *testing.T) {
-		s := newConnectionsState(nil, nil, routerconfig.Info{}, false)
+		s := newConnectionsState(nil, nil, balancerConfig.Info{}, false)
 		c, failed := s.GetConnection(context.Background())
 		require.Nil(t, c)
 		require.Equal(t, 0, failed)
@@ -370,7 +370,7 @@ func TestConnection(t *testing.T) {
 		s := newConnectionsState([]conn.Conn{
 			&mock.Conn{AddrField: "1", State: conn.Online},
 			&mock.Conn{AddrField: "2", State: conn.Online},
-		}, nil, routerconfig.Info{}, false)
+		}, nil, balancerConfig.Info{}, false)
 		c, failed := s.GetConnection(context.Background())
 		require.NotNil(t, c)
 		require.Equal(t, 0, failed)
@@ -379,7 +379,7 @@ func TestConnection(t *testing.T) {
 		s := newConnectionsState([]conn.Conn{
 			&mock.Conn{AddrField: "1", State: conn.Online},
 			&mock.Conn{AddrField: "2", State: conn.Banned},
-		}, nil, routerconfig.Info{}, false)
+		}, nil, balancerConfig.Info{}, false)
 		c, _ := s.GetConnection(context.Background())
 		require.Equal(t, &mock.Conn{AddrField: "1", State: conn.Online}, c)
 	})
@@ -387,9 +387,9 @@ func TestConnection(t *testing.T) {
 		s := newConnectionsState([]conn.Conn{
 			&mock.Conn{AddrField: "t1", State: conn.Banned, LocationField: "t"},
 			&mock.Conn{AddrField: "f2", State: conn.Banned, LocationField: "f"},
-		}, func(routerInfo routerconfig.Info, c conn.Conn) bool {
-			return c.Endpoint().Location() == routerInfo.SelfLocation
-		}, routerconfig.Info{}, true)
+		}, func(info balancerConfig.Info, c conn.Conn) bool {
+			return c.Endpoint().Location() == info.SelfLocation
+		}, balancerConfig.Info{}, true)
 		preferred := 0
 		fallback := 0
 		for i := 0; i < 100; i++ {
@@ -410,9 +410,9 @@ func TestConnection(t *testing.T) {
 		s := newConnectionsState([]conn.Conn{
 			&mock.Conn{AddrField: "t1", State: conn.Banned, LocationField: "t"},
 			&mock.Conn{AddrField: "f2", State: conn.Online, LocationField: "f"},
-		}, func(routerInfo routerconfig.Info, c conn.Conn) bool {
-			return c.Endpoint().Location() == routerInfo.SelfLocation
-		}, routerconfig.Info{SelfLocation: "t"}, true)
+		}, func(info balancerConfig.Info, c conn.Conn) bool {
+			return c.Endpoint().Location() == info.SelfLocation
+		}, balancerConfig.Info{SelfLocation: "t"}, true)
 		c, failed := s.GetConnection(context.Background())
 		require.Equal(t, &mock.Conn{AddrField: "f2", State: conn.Online, LocationField: "f"}, c)
 		require.Equal(t, 1, failed)
@@ -421,7 +421,7 @@ func TestConnection(t *testing.T) {
 		s := newConnectionsState([]conn.Conn{
 			&mock.Conn{AddrField: "1", State: conn.Online, NodeIDField: 1},
 			&mock.Conn{AddrField: "2", State: conn.Online, NodeIDField: 2},
-		}, nil, routerconfig.Info{}, false)
+		}, nil, balancerConfig.Info{}, false)
 		c, failed := s.GetConnection(WithEndpoint(context.Background(), &mock.Endpoint{AddrField: "2", NodeIDField: 2}))
 		require.Equal(t, &mock.Conn{AddrField: "2", State: conn.Online, NodeIDField: 2}, c)
 		require.Equal(t, 0, failed)
@@ -430,7 +430,7 @@ func TestConnection(t *testing.T) {
 		s := newConnectionsState([]conn.Conn{
 			&mock.Conn{AddrField: "1", State: conn.Online, NodeIDField: 1},
 			&mock.Conn{AddrField: "2", State: conn.Unknown, NodeIDField: 2},
-		}, nil, routerconfig.Info{}, false)
+		}, nil, balancerConfig.Info{}, false)
 		c, failed := s.GetConnection(WithEndpoint(context.Background(), &mock.Endpoint{AddrField: "2", NodeIDField: 2}))
 		require.Equal(t, &mock.Conn{AddrField: "1", State: conn.Online, NodeIDField: 1}, c)
 		require.Equal(t, 0, failed)