+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fixed data race on session stream queries
+* Renamed `internal/router` package to `internal/balancer` for unambiguous understanding of package mission
 * Implemented detection of local data-center with measuring tcp dial RTT
 * Added `trace.Driver.OnBalancer{Init,Close,ChooseEndpoint,Update}` events
 * Marked the driver cluster events as deprecated